requirements.txt
transformers
accelerate
pyannote-audio>=3.1.0
setuptools>=68.2.2
rich>=13.7.0
python-dotenv